// src/lib/client/api.ts
export interface Constraint {
    /**
     * 
     * @type {string}
     * @memberof Constraint
     */
    'message': string;
    /**
     * 
     * @type {string}
     * @memberof Constraint
     */
    'condition': string;
}
// ...
export interface Metadata {
    /**
     * 
     * @type {string}
     * @memberof Metadata
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof Metadata
     */
    'type': string;
    /**
     * 
     * @type {string}
     * @memberof Metadata
     */
    'entity'?: string;
    /**
     * 
     * @type {Array<Metadata>}
     * @memberof Metadata
     */
    'children'?: Array<Metadata>;
    /**
     * 
     * @type {Array<Property>}
     * @memberof Metadata
     */
    'fields'?: Array<Property>;
    /**
     * 
     * @type {string}
     * @memberof Metadata
     */
    'property'?: string;
    /**
     * 
     * @type {Array<Parameter>}
     * @memberof Metadata
     */
    'parameters'?: Array<Parameter>;
    /**
     * 
     * @type {Relation}
     * @memberof Metadata
     */
    'relation'?: Relation;
    /**
     * 
     * @type {string}
     * @memberof Metadata
     */
    'sentence'?: string;
    /**
     * 
     * @type {string}
     * @memberof Metadata
     */
    'alias'?: string;
    /**
     * 
     * @type {number}
     * @memberof Metadata
     */
    'number'?: number;
}
/**
 * 
 * @export
 * @interface MetadataConstraint
 */
export interface MetadataConstraint {
    /**
     * 
     * @type {string}
     * @memberof MetadataConstraint
     */
    'entity': string;
    /**
     * 
     * @type {Array<Constraint>}
     * @memberof MetadataConstraint
     */
    'constraints': Array<Constraint>;
    /**
     * 
     * @type {Array<MetadataConstraint>}
     * @memberof MetadataConstraint
     */
    'children'?: Array<MetadataConstraint>;
}
/**
 * 
 * @export
 * @interface MetadataModel
 */
export interface MetadataModel {
    /**
     * 
     * @type {string}
     * @memberof MetadataModel
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof MetadataModel
     */
    'type': string;
    /**
     * 
     * @type {Array<MetadataModel>}
     * @memberof MetadataModel
     */
    'children'?: Array<MetadataModel>;
}
/**
 * 
 * @export
 * @interface MetadataParameter
 */
export interface MetadataParameter {
    /**
     * 
     * @type {string}
     * @memberof MetadataParameter
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof MetadataParameter
     */
    'type': string;
    /**
     * 
     * @type {Array<MetadataParameter>}
     * @memberof MetadataParameter
     */
    'children'?: Array<MetadataParameter>;
}
/**
 * 
 * @export
 * @interface MetadataSentence
 */
export interface MetadataSentence {
    /**
     * 
     * @type {string}
     * @memberof MetadataSentence
     */
    'entity': string;
    /**
     * 
     * @type {string}
     * @memberof MetadataSentence
     */
    'dialect': string;
    /**
     * 
     * @type {string}
     * @memberof MetadataSentence
     */
    'dataSource': string;
    /**
     * 
     * @type {string}
     * @memberof MetadataSentence
     */
    'sentence': string;
    /**
     * 
     * @type {Array<MetadataSentence>}
     * @memberof MetadataSentence
     */
    'children'?: Array<MetadataSentence>;
}
/**
 * 
 * @export
 * @interface Parameter
 */
export interface Parameter {
    /**
     * 
     * @type {string}
     * @memberof Parameter
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof Parameter
     */
    'type': string;
    /**
     * 
     * @type {any}
     * @memberof Parameter
     */
    'value'?: any;
}
/**
 * 
 * @export
 * @interface Ping
 */
export interface Ping {
    /**
     * 
     * @type {string}
     * @memberof Ping
     */
    'message': string;
    /**
     * 
     * @type {string}
     * @memberof Ping
     */
    'time': string;
}
/**
 * 
 * @export
 * @interface Property
 */
export interface Property {
    /**
     * 
     * @type {string}
     * @memberof Property
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof Property
     */
    'type': string;
    /**
     * 
     * @type {number}
     * @memberof Property
     */
    'length'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof Property
     */
    'nullable'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Property
     */
    'primaryKey'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Property
     */
    'autoincrement'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof Property
     */
    'default'?: string;
    /**
     * 
     * @type {string}
     * @memberof Property
     */
    'readValue'?: string;
    /**
     * 
     * @type {string}
     * @memberof Property
     */
    'writeValue'?: string;
    /**
     * 
     * @type {string}
     * @memberof Property
     */
    'enum'?: string;
}
// ...
export interface QueryOptions {
    /**
     * 
     * @type {string}
     * @memberof QueryOptions
     */
    'stage'?: string;
    /**
     * 
     * @type {string}
     * @memberof QueryOptions
     */
    'view'?: string;
    /**
     * 
     * @type {number}
     * @memberof QueryOptions
     */
    'chunkSize'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof QueryOptions
     */
    'tryAllCan'?: boolean;
}
/**
 * 
 * @export
 * @interface QueryRequest
 */
export interface QueryRequest {
    /**
     * 
     * @type {QueryOptions}
     * @memberof QueryRequest
     */
    'options'?: QueryOptions;
    /**
     * 
     * @type {string}
     * @memberof QueryRequest
     */
    'expression': string;
    /**
     * 
     * @type {any}
     * @memberof QueryRequest
     */
    'data'?: any;
}
/**
 * 
 * @export
 * @interface QueryRequest
 */
export interface QueryQueuedRequest {
    /**
     * 
     * @type {QueryOptions}
     * @memberof QueryQueuedRequest
     */
    'options'?: QueryOptions;
    /**
     * 
     * @type {string}
     * @memberof QueryQueuedRequest
     */
    'expression': string;
    /**
     * 
     * @type {any}
     * @memberof QueryQueuedRequest
     */
    'data'?: any;
     /**
     * 
     * @type {string}
     * @memberof QueryQueuedRequest
     */
     'topic': string;
     /**
     * 
     * @type {number}
     * @memberof QueryQueuedRequest
     */
     'chunk'?: number;
}
/**
 * 
 * @export
 * @interface QueryResponse
 */
export interface QueryResponse {
    /**
     * 
     * @type {any}
     * @memberof QueryResponse
     */
    'result': any;
}
/**
 * 
 * @export
 * @interface Relation
 */
export interface Relation {
    /**
     * 
     * @type {string}
     * @memberof Relation
     */
    'name': string;
    /**
     * 
     * @type {RelationType}
     * @memberof Relation
     */
    'type': RelationType;
    /**
     * 
     * @type {boolean}
     * @memberof Relation
     */
    'composite'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof Relation
     */
    'from': string;
    /**
     * 
     * @type {string}
     * @memberof Relation
     */
    'entity': string;
    /**
     * 
     * @type {string}
     * @memberof Relation
     */
    'to': string;
    /**
     * 
     * @type {boolean}
     * @memberof Relation
     */
    'weak'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof Relation
     */
    'target'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Relation
     */
    'targetComposite'?: boolean;
}
/**
 * 
 * @export
 * @enum {string}
 */

export const RelationType = {
    OneToMany: 'oneToMany',
    ManyToOne: 'manyToOne',
    OneToOne: 'oneToOne'
} as const;

export type RelationType = typeof RelationType[keyof typeof RelationType];
// ...
export class ExpressionApi extends BaseAPI {
    /**
     * 
     * @param {QueryRequest} [queryRequest] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ExpressionApi
     */
    public constraints(queryRequest?: QueryRequest, options?: AxiosRequestConfig) {
        return ExpressionApiFp(this.configuration).constraints(queryRequest, options).then((request) => request(this.axios, this.configuration.basePath))
    }

    /**
     * 
     * @param {QueryRequest} queryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ExpressionApi
     */
    public execute(queryRequest: QueryRequest, options?: AxiosRequestConfig) {
        return ExpressionApiFp(this.configuration).execute(queryRequest, options).then((request) => request(this.axios, this.configuration.basePath))
    }

    /**
     * 
     * @param {QueryQueuedRequest} queryQueuedRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ExpressionApi
     */
    public executeQueued(queryQueuedRequest: QueryQueuedRequest, options?: AxiosRequestConfig) {
        return ExpressionApiFp(this.configuration).execute(queryQueuedRequest, options).then((request) => request(this.axios, this.configuration.basePath))
    }

    /**
     * 
     * @param {QueryRequest} [queryRequest] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ExpressionApi
     */
    public metadata(queryRequest?: QueryRequest, options?: AxiosRequestConfig) {
        return ExpressionApiFp(this.configuration).metadata(queryRequest, options).then((request) => request(this.axios, this.configuration.basePath))
    }

    /**
     * 
     * @param {QueryRequest} [queryRequest] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ExpressionApi
     */
    public model(queryRequest?: QueryRequest, options?: AxiosRequestConfig) {
        return ExpressionApiFp(this.configuration).model(queryRequest, options).then((request) => request(this.axios, this.configuration.basePath))
    }

    /**
     * 
     * @param {QueryRequest} [queryRequest] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ExpressionApi
     */
    public parameters(queryRequest?: QueryRequest, options?: AxiosRequestConfig) {
        return ExpressionApiFp(this.configuration).parameters(queryRequest, options).then((request) => request(this.axios, this.configuration.basePath))
    }

    /**
     * 
     * @param {QueryRequest} queryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ExpressionApi
     */
    public sentence(queryRequest: QueryRequest, options?: AxiosRequestConfig) {
        return ExpressionApiFp(this.configuration).sentence(queryRequest, options).then((request) => request(this.axios, this.configuration.basePath))
    }
}